Add current progress of SoC
