+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "workbench.editor.enablePreview": false,
+    "workbench.editor.wrapTabs": true
+}
